usuario premium y cambiar contresaña

+++ src/controllers/product.controller.js
@@ -7,7 +7,6 @@ import logger from '../logger.js'
 export const createProductController = async (req, res) => {
   try {
     const product = req.body
-    // console.log(product)
     if (!product.title || !product.price || !product.description || !product.code || !product.stock || !product.category) {
       const errorInfo = generateProductErrorInfo(product);
       CustomError.createError({
@@ -63,6 +62,20 @@ export const updateProductController = async (req, res) => {
 export const deleteProductController = async (req, res) => {
   try {
     const productId = req.params.pid;
+
+    const product = await ProductService.getById(productId)
+    const userInfo = {
+      first_name: req.session.user.first_name,
+      last_name: req.session.user.last_name,
+      email: req.session.user.email,
+      age: req.session.user.age,
+      role: req.session.user.role,
+    };
+
+    if (product.owner != userInfo.email && userInfo.role == 'premium'){
+      res.status(404).json({ error: 'No puedes eliminar este producto' })
+      return;
+    }
 
     //const deletedProduct = await Product.findByIdAndDelete(productId).lean().exec();
     const deletedProduct = await ProductService.delete(productId)
